requirements.txt
autopep8==2.0.4
EasyProcess==1.1
entrypoint2==1.1
Levenshtein==0.23.0
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
pycodestyle==2.11.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-Levenshtein==0.23.0
pytweening==1.0.7
rapidfuzz==3.4.0
six==1.16.0
tomli==2.0.1
